cleanup the claims mapping to its own pkg
