+++ providers/oidc.go
@@ -6,16 +6,12 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
-	"math"
 	"net/http"
-	"reflect"
-	//"sort"
 	"time"
 
 	oidc "github.com/coreos/go-oidc"
-	//"github.com/davecgh/go-spew/spew"
-	//"github.com/mozillazg/go-httpheader"
 	"github.com/pusher/oauth2_proxy/pkg/apis/sessions"
+	claimsconv "github.com/pusher/oauth2_proxy/pkg/claims"
 	"golang.org/x/oauth2"
 )
 
@@ -204,95 +200,6 @@ func (p *OIDCProvider) ValidateSessionState(s *sessions.SessionState) bool {
 	return true
 }
 
-// valueString returns the string representation of a value.
-func valueString(v reflect.Value) string {
-	for v.Kind() == reflect.Ptr {
-		if v.IsNil() {
-			return ""
-		}
-		v = v.Elem()
-	}
-
-	if v.Kind() == reflect.Bool {
-		if v.Bool() {
-			return "1"
-		}
-		return "0"
-	}
-
-	if v.Type() == reflect.TypeOf(time.Time{}) {
-		t := v.Interface().(time.Time)
-		return t.Format(http.TimeFormat)
-	}
-
-	return fmt.Sprint(v.Interface())
-}
-
-func reflectValue(header http.Header, val reflect.Value) error {
-	var embedded []reflect.Value
-
-	iter := val.MapRange()
-	for iter.Next() {
-		name := iter.Key().String()
-		sv := iter.Value()
-
-		switch t := sv.Interface().(type) {
-		case int:
-			//log.Printf("%s => %s", name, string(t))
-			header.Add(name, string(t))
-			continue
-		case float64:
-			if name == "exp" || name == "nbf" || name == "iat" {
-				// well-known time entries
-				sec, dec := math.Modf(t)
-				vv := time.Unix(int64(sec), int64(dec*1e9))
-				header.Add(name, vv.Format(http.TimeFormat))
-			} else {
-				header.Add(name, fmt.Sprintf("%f", t))
-			}
-			continue
-		case string:
-			//log.Printf("%s => %v", name, t)
-			header.Add(name, t)
-			continue
-		case bool:
-			//log.Printf("%s => %v", name, t)
-			if t {
-				header.Add(name, "true")
-			} else {
-				header.Add(name, "false")
-			}
-			continue
-		case []interface{}: // array
-			for _, item := range t {
-				vv := valueString(reflect.ValueOf(item))
-				//log.Printf("%s[%d] ==> %s", name, i, vv)
-				header.Add(name, vv)
-			}
-			continue
-		case map[string]interface{}: // map
-			//log.Printf("%s => %v", name, t)
-			for key, item := range t {
-				vv := valueString(reflect.ValueOf(item))
-				//log.Printf("%s[%s] ==> %s", name, key, vv)
-				header.Add(fmt.Sprintf("%s_%s", name, key), vv)
-			}
-			continue
-		default:
-			log.Printf("%s unknown cast => %v", name, t)
-			continue
-		}
-	}
-
-	for _, f := range embedded {
-		if err := reflectValue(header, f); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 // HeadersToInject processes the OIDC session and publishes a set of headers to export to upstreams
 func (p *OIDCProvider) HeadersToInject(s *sessions.SessionState) (*http.Header, error) {
 	ctx := context.Background()
@@ -311,8 +218,7 @@ func (p *OIDCProvider) HeadersToInject(s *sessions.SessionState) (*http.Header, 
 	var idClaims map[string]interface{}
 	token.Claims(&idClaims)
 	h := http.Header{}
-	val := reflect.ValueOf(idClaims)
-	reflectValue(h, val)
+	claimsconv.EncodeHeaders(h, idClaims)
 	//h, err := httpheader.Header(idClaims)
 	//if err != nil {
 	//	return nil, fmt.Errorf("Unable to serialize id claims: %v", err)
@@ -324,9 +230,8 @@ func (p *OIDCProvider) HeadersToInject(s *sessions.SessionState) (*http.Header, 
 	if err != nil {
 		return nil, fmt.Errorf("Unable to parse userinfo: %v", err)
 	}
-	userVal := reflect.ValueOf(userClaims)
 	hU := http.Header{}
-	reflectValue(hU, userVal)
+	claimsconv.EncodeHeaders(hU, userClaims)
 
 	// Merge the maps
 	for k, v := range hU {
